The prefpane works.
This means that we're almost ready to ship -- we still need feature parity and new features, but We Work Well at the moment :D

Also commited some art. May be wise to revise early Afloat 1.0 web site/marketing sketches to reflect the new icon art (white on blue with gradients).